Split header into title, byline and description elements and created spacer elements for better padding control.
It used to be controlled by a combination of `font-size`, `line-height` and `height`, which is a horrible way to control padding. With the `line-height` and `height` set equal to `font-size`, padding between elements can now be controlled by the actual `padding` property! Which makes tuning these things a lot less convoluted and easier to manage.

Also set ``#yara-rule-designer-theme-selector` position to absolute to force the buttons to be at top left withour eating a whole row of space from the header (or any other) element.